requirements.txt
fastai==2.5.3
gradio==3.0.20
lightgbm==3.2.1
pandas==1.3.3
numpy==1.21.2
torch==1.9.0
torchvision==0.10.0
scikit-learn==0.24.2
imageio
pydicom
pillow